different headers in the snuba and kafka backends
