+++ src/sentry/eventstream/snuba.py
@@ -6,7 +6,7 @@ from uuid import uuid4
 import pytz
 import urllib3
 
-from sentry import options, quotas
+from sentry import quotas
 from sentry.eventstream.base import EventStream
 from sentry.utils import json, snuba
 from sentry.utils.safe import get_path
@@ -74,6 +74,19 @@ class SnubaProtocolEventStream(EventStream):
     # non-prefixed variations occur in a response.
     UNEXPECTED_TAG_KEYS = frozenset(["dist", "release", "user"])
 
+    def _get_headers_for_insert(
+        self,
+        group,
+        event,
+        is_new,
+        is_regression,
+        is_new_group_environment,
+        primary_hash,
+        received_timestamp,  # type: float
+        skip_consume,
+    ) -> Mapping[str, str]:
+        return {"Received-Timestamp": str(received_timestamp)}
+
     def insert(
         self,
         group,
@@ -99,43 +112,16 @@ class SnubaProtocolEventStream(EventStream):
         if unexpected_tags:
             logger.error("%r received unexpected tags: %r", self, unexpected_tags)
 
-        # HACK: We are putting all this extra information that is required by the
-        # post process forwarder into the headers so we can skip parsing entire json
-        # messages. The post process forwarder is currently bound to a single core.
-        # Once we are able to parallelize the JSON parsing and other transformation
-        # steps being done there we may want to remove this hack.
-        def encode_bool(value: Optional[bool]) -> str:
-            if value is None:
-                value = False
-            return str(int(value))
-
-        # WARNING: We must remove all None headers. There is a bug in confluent-kafka-python
-        # (used by both Sentry and Snuba) that incorrectly decrements the reference count of
-        # Python's None on any attempt to read header values containing null values, leading
-        # None to eventually get deallocated and crash the interpreter.
-        def strip_none_values(value: Mapping[str, Optional[str]]) -> Mapping[str, str]:
-            return {key: value for key, value in value.items() if value is not None}
-
-        send_new_headers = options.get("eventstream:kafka-headers")
-
-        if send_new_headers is True:
-            headers = strip_none_values(
-                {
-                    "Received-Timestamp": str(received_timestamp),
-                    "event_id": str(event.event_id),
-                    "project_id": str(event.project_id),
-                    "group_id": str(event.group_id) if event.group_id is not None else None,
-                    "primary_hash": str(primary_hash) if primary_hash is not None else None,
-                    "is_new": encode_bool(is_new),
-                    "is_new_group_environment": encode_bool(is_new_group_environment),
-                    "is_regression": encode_bool(is_regression),
-                    "version": str(self.EVENT_PROTOCOL_VERSION),
-                    "operation": "insert",
-                    "skip_consume": encode_bool(skip_consume),
-                }
-            )
-        else:
-            headers = {"Received-Timestamp": str(received_timestamp)}
+        headers = self._get_headers_for_insert(
+            group,
+            event,
+            is_new,
+            is_regression,
+            is_new_group_environment,
+            primary_hash,
+            received_timestamp,
+            skip_consume,
+        )
 
         self._send(
             project.id,
